# 1.0 perf review