--- a/brjs-core/src/main/java/org/bladerunnerjs/api/TestPack.java
+++ b/brjs-core/src/main/java/org/bladerunnerjs/api/TestPack.java
@@ -43,7 +43,7 @@ public class TestPack extends AbstractBundlableNode implements NamedNode
 	@Override
 	public List<LinkedAsset> seedAssets() 
 	{		
-		return assetDiscoveryInitiator.seedAssets();
+		return new ArrayList<>( assetDiscoveryResult().getRegisteredSeedAssets() );
 	}
 	
 	@Override