--- a/brjs-core/src/main/java/org/bladerunnerjs/api/plugin/AssetRegistry.java
+++ b/brjs-core/src/main/java/org/bladerunnerjs/api/plugin/AssetRegistry.java
@@ -1,6 +1,7 @@
 package org.bladerunnerjs.api.plugin;
 
 import java.util.List;
+import java.util.Set;
 
 import org.bladerunnerjs.api.Asset;
 import org.bladerunnerjs.api.LinkedAsset;
@@ -15,5 +16,7 @@ public interface AssetRegistry
 	boolean hasSeedAsset(String requirePath);
 	boolean hasRegisteredAsset(String requirePath);
 	public Asset getRegisteredAsset(String requirePath);
+	public Set<Asset> getRegisteredAssets();
+	public Set<LinkedAsset> getRegisteredSeedAssets();
 	public List<Asset> discoverFurtherAssets(MemoizedFile dir, String requirePrefix, List<Asset> implicitDependencies);
 }