… brjs-core/src/main/java/org/bladerunnerjs/api/plugin/AssetContainerAssets.java
… brjs-core/src/main/java/org/bladerunnerjs/utility/TailBuffer.java
… brjs-core/src/test/java/org/bladerunnerjs/utility/TailBufferTest.java