Performance improvements following the 1.0-RC1 release. This should fix #1413 and, in my tests, reduced fxtrader initial load time from 13s to 5s.

- The current performance of latest `develop` is already better than `1.0-RC1`, this PR improves it further
- Replaced `TailBuffer` which all our reader filters use with a more performant fixed length StringBuilder
- Replacing the regex we had in `JsModuleExportsStrippingReader` with a simple `.contains()` check

Based on my testing the times are:

### master (bf891cd458c9916e8394f54ca67c91732b19c341)

#### all BRJS JS tests

```
real	0m37.249s
user	1m22.478s
sys		0m4.423s
```

#### dashboard loading test

```
BRJS initialization: 265 ms
Cold Request: 1078 ms
Cold Request: 525 ms (all files modified but BRJS initialized)
Warm Request: 436 ms (some files modified)
Hot Request: 29 ms
```

#### dashboard loading test w/ fxtrader libraries

```
BRJS initialization: 307 ms
Cold Request: 1694 ms
Cold Request: 902 ms (all files modified but BRJS initialized)
Warm Request: 644 ms (some files modified)
Hot Request: 63 ms
```

#### fxtrader loading test

```
BRJS initialization: 269 ms
Cold Request: 13870 ms
Cold Request: 9296 ms (all files modified but BRJS initialized)
Warm Request: 8517 ms (some files modified)
Hot Request: 847 ms
```

#### pre 1.0-perf-review (e2239e0724384b25338a9cddb58a00ed8ba4f8d1)

#### All BRJS JS tests

```
real	0m34.957s
user	1m17.551s
sys		0m4.236s
```

#### dashboard loading test

```
BRJS initialization: 268 ms
Cold Request: 1010 ms
Cold Request: 493 ms (all files modified but BRJS initialized)
Warm Request: 457 ms (some files modified)
Hot Request: 27 ms
```

#### dashboard loading test w/ fxtrader libs

```
BRJS initialization: 311 ms
Cold Request: 1504 ms
Cold Request: 782 ms (all files modified but BRJS initialized)
Warm Request: 618 ms (some files modified)
Hot Request: 52 ms
```

#### fxtrader loading test

```
BRJS initialization: 267 ms
Cold Request: 9884 ms
Cold Request: 9156 ms (all files modified but BRJS initialized)
Warm Request: 8608 ms (some files modified)
Hot Request: 333 ms
```

### 1.0-perf-review branch (7e09b560ad0f6bcfb4c94d22e42afab50fd7d2a7):

#### All BRJS JS tests

```
real	0m33.996s
user	1m17.900s
sys		0m4.117s
```

#### dashboard loading test

```
BRJS initialization: 268 ms
Cold Request: 1049 ms
Cold Request: 493 ms (all files modified but BRJS initialized)
Warm Request: 439 ms (some files modified)
Hot Request: 28 ms
```

#### dashboard loading test w/ fxtrader libs

```
BRJS initialization: 335 ms
Cold Request: 1447 ms
Cold Request: 591 ms (all files modified but BRJS initialized)
Warm Request: 440 ms (some files modified)
Hot Request: 29 ms
```

#### fxtrader loading test

```
BRJS initialization: 262 ms
Cold Request: 5591 ms
Cold Request: 5237 ms (all files modified but BRJS initialized)
Warm Request: 4594 ms (some files modified)
Hot Request: 143 ms
```


<!---
@huboard:{"order":706.5,"milestone_order":1420,"custom_state":""}
-->