@thanhc 2015-06-10
I've done some checks using the fx motif:

                                    0.15                  RC1.0             PerfImprovements
Login page first load     ~3 secs            ~12 secs                 ~8 secs
App first load               ~18 secs           ~30 secs                 ~22 secs

The times for page reload and login second time are roughly the same on all versions.
As you can see there are some performance improvements, but still slightly slower than 0.15 on first load.