@thanhc 2015-06-16
Did some timings on BRJS running test: time brjs test libs/javascript/br-presenter

on 0.15: 
real    0m12.913s
user    0m0.075s
sys     0m0.197s

on RC 1.0:
real    0m16.300s
user    0m0.090s
sys     0m0.181s

On this branch:
real    0m12.892s
user    0m0.060s
sys     0m0.198s